backend/requirements.txt
chromadb==1.0.9
fastapi==0.115.9
httpx==0.28.1
openai==1.68.2
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
scikit-learn==1.5.2
uvicorn==0.34.0
